fix message typo and a warning
git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@1356 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ Common/Stylization/SE_StyleVisitor.cpp
@@ -466,7 +466,7 @@ void SE_StyleVisitor::VisitImage(Image& image)
         char* src_ascii = new char[srclen];
 
         char* ptr = src_ascii;
-        for (int i=0; i<srclen; i++)
+        for (size_t i=0; i<srclen; i++)
             *ptr++ = (char)src_u[i];
                 
         size_t dstlen = Base64::GetDecodedLength(image.GetContent().size());
